backend/requirements.txt
adafruit-io==2.8.0
asgiref==3.8.1
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
firebase-admin==6.7.0
fonttools==4.56.0
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.1
msgpack==1.1.0
numpy==2.2.4
packaging==24.2
paho-mqtt==2.1.0
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
urllib3==2.3.0
